back/requirements.txt
fastapi
uvicorn
pydantic
requests
python-dotenv
httpx==0.23.3 #youtube-search-python 할 때 중요함
pillow
pytesseract
python-multipart